% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities around the world.}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{continent}{Continent}
\item{country}{Country}
\item{city}{city}
}
}
\source{
The world
}
\usage{
cities
}
\description{
Dataset with hierarchical levels for cities.
}
\keyword{datasets}
