% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{World's countries.}
\format{
A data frame with 177 rows and 3 variables:
\describe{
\item{continent}{Continent}
\item{subregion}{Sub region}
\item{name}{Country name}
}
}
\source{
Data from package \href{https://github.com/ropensci/rnaturalearth}{rnaturalearth}
}
\usage{
countries
}
\description{
Dataset with hierarchical levels for countries : continent and subregion.
}
\keyword{datasets}
