% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-treeview.R
\name{revealTreeviewSelected}
\alias{revealTreeviewSelected}
\title{Reveals selected node in \code{treeviewInput}}
\usage{
revealTreeviewSelected(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The id of the input object.}

\item{session}{The session object passed to function given to shinyServer.}
}
\value{
None.
}
\description{
Reveals selected node in \code{treeviewInput}
}
