% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-treeview.R
\name{make_tree}
\alias{make_tree}
\title{Create choice structure for \code{treeviewInput}}
\usage{
make_tree(data, levels)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{levels}{Variables identifying hierarchical levels.}
}
\value{
a \code{list} that can be used in \code{\link{treeviewInput}}.
}
\description{
Create choice structure for \code{treeviewInput}
}
\examples{
data("cities")
head(cities)

make_tree(cities, c("continent", "country", "city"))
}
