% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-treeview.R
\name{treeviewInput}
\alias{treeviewInput}
\title{Title}
\usage{
treeviewInput(
  inputId,
  label = NULL,
  choices,
  selected = NULL,
  multiple = FALSE,
  levels = 1,
  borders = TRUE,
  prevent_unselect = FALSE,
  return_value = c("name", "all"),
  width = NULL
)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
